Add code for plotting total number of SNe. Change graph properties. Update stats

+++ py/mk_star_stats.py
@@ -4,6 +4,7 @@ import urllib.request
 
 from beautifulsoup_supply import TAIL, mk_head, get_soup, get_soup_Request
 from plot_supply import plot_bar, optimize_svg
+import matplotlib.pyplot as plt
 
 
 HEAD = mk_head("Статистика звездных каталогов", style="../../compact.css", script="") + "<body>\n"
@@ -78,7 +79,7 @@ snstats = {}
 snstats_txt = f"""<h2><a href="{SNSTATS_URL}">Статистика вспышек сверхновых</a></h2>
 <ul>
 """
-years, sns, snalt = [], [], []
+years, sns, snalt, all_sne = [], [], [], []
 all_sn_count, sn_amateur_count = 0, 0
 for (year, snstats_url) in snurls:
     soup = get_soup(snstats_url)
@@ -87,11 +88,13 @@ for (year, snstats_url) in snurls:
     all_sn_count += sn_num
     sn_amateur_count += sn_amateur
     if year == 1995:
+        all_sne.append(all_sn_count)
         snstats_txt += f"<li><a href='{snstats_url}' target='_blank' rel='noopener noreferrer'>До 1996 года</a> открыто <b>{sn_num}</b> сверхновых, <b>{sn_amateur_count}</b> – любителями, <b>{sn_13th}</b> ярче 13-й зв. величины (<a href='{SNOTHER_URL}'>яркие сверхновые до 1996 года</a>).\n"
         years.append(year)
         sns.append(sn_num - sn_amateur)
         snalt.append(sn_amateur)
     elif year != "all":
+        all_sne.append(all_sn_count)
         snstats_txt += f"<li><a href='{snstats_url}' target='_blank' rel='noopener noreferrer'>За {year} год</a> открыто <b>{sn_num}</b> сверхновых, <b>{sn_amateur}</b> – любителями, <b>{sn_13th}</b> ярче 13-й зв. величины. Всего к концу года открыто <b>{all_sn_count}</b>, <b>{sn_amateur_count}</b> – любителями.\n"
         years.append(year)
         sns.append(sn_num - sn_amateur)
@@ -99,22 +102,47 @@ for (year, snstats_url) in snurls:
     else:
         snstats_txt += f"""<li><a href="{snstats_url}" target="_blank" rel="noopener noreferrer">Всего открыто</a> <b>{sn_num}</b> сверхновых, <b>{sn_amateur}</b> – любителями, <b>{sn_13th}</b> ярче 13-й зв. величины.\n"""
 
-labels = ('Статистика открытий сверхновых по годам', 'Год', 'Открытий СН за год')
-tmp_filename = 'snstats_plot_.svg'
-filename = 'snstats_plot.svg'
+file_ext = 'svg'
+total_sne_filename = 'sne_total_number_log_plot.' + file_ext
 stars_dir = os.path.join(os.pardir, 'stars')
+pth = os.path.join(stars_dir, total_sne_filename)
+tmp_filename = 'sne_total_number_log_plot_.' + file_ext
+tmp_pth = os.path.join(stars_dir, tmp_filename)
+
+fig, ax = plt.subplots(figsize=(16, 9))
+plt.subplots_adjust(left=0.06, bottom=0.06, right=0.97, top=0.955)
+
+plt.plot(years, all_sne, 'ok-')
+plt.yscale("log")
+ax.minorticks_on()
+plt.title('Динамика открытий сверхновых', fontsize=16)
+plt.xlabel('Время', fontsize=14)
+plt.ylabel('Количество открытых сверхновых', fontsize=14)
+plt.grid(which='major', linestyle='-')
+plt.grid(which='minor', linestyle='--')
+plt.savefig(tmp_pth, dpi=240)
+if file_ext == 'svg':
+    optimize_svg(tmp_pth, pth)
+    os.remove(tmp_pth)
+
+
+labels = ('Статистика открытий сверхновых по годам', 'Год',
+          'Открытий сверхновых за год')
+tmp_filename = 'snstats_plot_.' + file_ext
+filename = 'snstats_plot.' + file_ext
 tmp_pth = os.path.join(stars_dir, tmp_filename)
 pth = os.path.join(stars_dir, filename)
 xlim = (1994.3, 2022.7)
 plot_bar(years, sns, snalt, labels, tmp_pth, xlim, lab0="до 1996")
-optimize_svg(tmp_pth, pth)
-os.remove(tmp_pth)
+if file_ext == 'svg':
+    optimize_svg(tmp_pth, pth)
+    os.remove(tmp_pth)
 soup = get_soup_Request(TNS_STATS_URL)
 all_transient, public_transient, classified, spectra = get_tns(soup)
 
 snstats_txt += f"""</ul>
-<br><img src="../{filename}" alt="">
-
+<br><img src="../{filename}" alt=""><br>
+<img src="../{total_sne_filename}" alt="">
 <h2><a href="{TNS_URL}">Transient Name Server</a></h2>
 <a href="{TNS_STATS_URL}" target="_blank" rel="noopener noreferrer">статистика</a>:<br>
 <ul>
